[libc++] Split features.py into multiple files
The features.py file that performs detection of Lit features had grown
to be massive, so this patch splits it into smaller chunks which makes
it easier to keep things organized.
